Refactor BaseLLM
